fix: fix missing vibrant color on iOS
